cli(deploy): one module per host, and a dispatcher
runDeploy was a 300-line function with four host branches inline, each
carrying its host's kept-artifact semantics and gates, followed by a
runDeployX wrapper per host and a cross-host isOurArtifact table. A host
is now cli/commands/deploy/<host>.ts exporting a HostDeploy — isOurs and
deploy(ctx) — over the neutral deploy/<host>/ plan + driver; deploy.ts
is the dispatcher (prelude, preflight, the host-only-flag table, HOSTS).
The bodies moved verbatim; writeArtifacts asks the host which files are
its own and answers for the container's two itself.
