+++ test/deploy-artifacts.test.ts
@@ -12,7 +12,7 @@ import { mkdtemp, mkdir, readFile, writeFile } from "node:fs/promises";
 import { dirname, join } from "node:path";
 import { tmpdir } from "node:os";
 import { describe, expect, it, vi } from "vitest";
-import { HOST_ONLY_FLAGS, warnHostOnlyFlags, writeArtifacts } from "../src/cli/commands/deploy.ts";
+import { HOSTS, HOST_ONLY_FLAGS, warnHostOnlyFlags, writeArtifacts } from "../src/cli/commands/deploy.ts";
 import { DEPLOY_HOSTS } from "../src/deploy/hosts.ts";
 
 /** The opening bytes that make each artifact kind OURS — one per kind the ownership table knows. */
@@ -25,6 +25,9 @@ const MARKERS = {
   "fastagent/agentcore.template.yaml": "# Generated by `fastagent deploy agentcore`",
 } as const;
 type Kind = keyof typeof MARKERS;
+
+/** Every host's answer to "is this yours?" — the table the ownership rule is made of. */
+const isOurs = (path: string, content: string): boolean => Object.values(HOSTS).some((h) => h.isOurs(path, content));
 
 /** Generated content for `path`: the marker must OPEN the file, which is the rule being tested. */
 const generatedFor = (path: Kind): string =>
@@ -44,7 +47,7 @@ async function attempt(
   const said: string[] = [];
   const spy = vi.spyOn(console, "error").mockImplementation((m: unknown) => said.push(String(m)));
   try {
-    await writeArtifacts(dir, [{ path, content: generatedFor(path) }], { force });
+    await writeArtifacts(dir, [{ path, content: generatedFor(path) }], { force, isOurs });
   } finally {
     spy.mockRestore();
   }
@@ -74,7 +77,7 @@ describe("deploy: ownership decides what --force may replace", () => {
     const said: string[] = [];
     const spy = vi.spyOn(console, "error").mockImplementation((m: unknown) => said.push(String(m)));
     try {
-      await writeArtifacts(dir, [{ path, content: "// fresh\n" }], { force: false, alwaysWrite: [path] });
+      await writeArtifacts(dir, [{ path, content: "// fresh\n" }], { force: false, alwaysWrite: [path], isOurs });
     } finally {
       spy.mockRestore();
     }
